backend/api/requirements.txt
tensorflow==2.15.0
tensorflow-serving-api==2.14.1
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
pillow==10.2.0
matplotlib==3.8.3
numpy==1.26.4
torch
torchvision
torchaudio
